Create API.py
To automate the collection of location data from Google Maps, you can use the Google Maps API. Here's a basic overview of how you can use the API to collect location data:

Get an API key: To access the Google Maps API, you'll need to obtain an API key from the Google Cloud Console. The API key is used to authenticate your request and track usage of the API.

Choose an API endpoint: The Google Maps API offers several endpoints for retrieving different types of data, such as directions, geocoding, and places. Choose the endpoint that matches the data you want to collect. For location data, you can use the Places API.

Make an API request: Once you have an API key and have chosen an endpoint, you can make an API request by sending a GET request to the API endpoint, along with your API key and any parameters required to specify the data you want to retrieve.

Parse the API response: The API will respond with a JSON object that contains the location data you requested. You'll need to parse the response in order to extract the information you want to collect.

Automate the process: Once you have the code to make an API request and parse the response, you can automate the process by scheduling the code to run at regular intervals, or by triggering the code in response to specific events.

Here's an example of what the code might look like in Python:

This code makes a text search request to the Google Maps Places API, asking for locations with the name "Emart". The response is a JSON object that contains information about the location, including its name, address, latitude, and longitude. The code parses the response and prints the information for each location.